# specs/001-todo-app-fullstack/contracts/task-api-contract.yaml
openapi: 3.0.3
info:
  title: Todo App Task API
  description: API for managing user tasks in the Todo Web Application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Retrieve all tasks associated with the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user whose tasks to retrieve
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, completed, incomplete]
          description: Filter tasks by completion status
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this user's tasks
        '404':
          description: User not found
      security:
        - bearerAuth: []

    post:
      summary: Create a new task
      description: Create a new task for the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to create the task for
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to create tasks for this user
      security:
        - bearerAuth: []

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by its ID and user ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to retrieve
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this task
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    put:
      summary: Update a task
      description: Update an existing task for the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to update this task
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    delete:
      summary: Delete a task
      description: Delete an existing task for the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to delete this task
        '404':
          description: Task not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the task
        user_id:
          type: string
          description: Reference to the user who owns the task
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          nullable: true
          description: Detailed task description
        completed:
          type: boolean
          description: Task completion status
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Detailed task description
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Detailed task description
        completed:
          type: boolean
          description: Task completion status
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task